Possible fix for file like objects
In current hickle file like objects are closed and replaced by native
hickle file. This poses problems when hickle file should be stored in
ByteIO or other readable, writeable and seekable file and file like
objects or read from readable and seekable file and file like object.
The proposed fix checks for readable and seekable method defined by
IOBase for file to be read and in additon for writeable if file should
be written or appended. Objects which do not define these three method
but provide read, seek, tell and in addition write for writing than file
or file like object is passed to h5py.File instead of file name.
Accepts following mode specifiers 'w', 'w+', 'r', 'r+', 'x', 'x+' and
'a'. According to h5py manual passed in file or file like object must be
in binary mode, means write accepts bytes, bytearray or alike and read
returns bytes or bytearray. A trailing 'b' is removed from mode before
passed to h5py.File. This allows to use the same mode specifier as
passed to open. But carefull file and file like object to be written must
be opened for reading and writing. With open use 'w+b' instead of 'wb'
or 'x+b' instead of 'xb'.